moved old settings page to data integration page within new setting page for tennant
