+++ packages/apps/spaces/app/settings/Tabs/panels/AuthPanel/index.ts
@@ -1,1 +1,1 @@
-export * from './OAuthPanel';
+export * from './AuthPanel';
