+++ packages/apps/spaces/app/settings/Tabs/panels/IntegrationsPanel/index.ts
@@ -1,0 +1,1 @@
+export * from './IntegrationsPanel';
