--- packages/apps/spaces/pages/integrations/integrations.module.scss
@@ -1,11 +1,0 @@
-.pageContainer {
-}
-
-.headingSection {
-  margin-bottom: var(--spacing-sm);
-}
-
-.settingsContainer {
-  display: flex;
-  flex-direction: column;
-}
